requirements.txt
python==3.10.11
paddlepaddle==2.6.2
paddleocr==2.7.0.3
numpy==1.26.4
opencv-python==4.6.0.66
torch==2.0.1+cpu
transformers==4.37.2
sentencepiece
streamlit
pandas
Pillow
langdetect